Revert "Remove withRouter from UploadAirgapBundle (#3343)"
This reverts commit b43fa9a89fe33bddb3a32edfd5b9fe415324bb97.
